[refinement] improved refinement limits implied by 32-bit integers in 2D (Morton_id limit remains)

+++ src/grid/refinement_flag.F90
@@ -184,7 +184,7 @@ contains
       implicit none
 
       class(ref_flag_t), intent(inout) :: this     !> object invoking this procedure
-      integer(kind=4), intent(in)      :: il, jl, kl, ih, jh, kh  !> box coordinates
+      integer(kind=8), intent(in)      :: il, jl, kl, ih, jh, kh  !> box coordinates
       logical, dimension(il:ih, jl:jh, kl:kh), optional, intent(in) :: mask  !> the mask to filter out something
 
 #ifdef DEBUG
